// src/app/posts/models/institution.ts
export interface Institution {
    uuid:           string;
    name:           string;
    description:    string;
    location:       string;
    category:       string;
    email:          string;
    phone:          string;
    url:            string;
    logo_url:       string;
    background_url: string;
}